backend/requirements.txt
fastapi
uvicorn
chromadb
google-generativeai
pydantic
python-dotenv
requests
beautifulsoup4
